requirements.txt
tweepy==4.14.0
anthropic==0.42.0
pyyaml==6.0.2
python-dotenv==1.0.1
schedule==1.2.2
pytz==2024.2
